[bcr-wdc-treasury-service] move crsat into foreign
for future use from incoming sat module
