+++ crates/bcr-wdc-treasury-service/src/web.rs
@@ -4,7 +4,7 @@ use bcr_wdc_webapi::exchange as web_exchange;
 use cashu::nut03 as cdk03;
 // ----- extra library imports
 // ----- local imports
-use crate::{crsat, debit, error::Result};
+use crate::{foreign, debit, error::Result};
 
 #[tracing::instrument(level = tracing::Level::DEBUG, skip(ctrl))]
 pub async fn redeem<Wlt, WdcSrvc, Repo>(
@@ -25,7 +25,7 @@ where
 #[tracing::instrument(level = tracing::Level::DEBUG, skip(ctrl))]
 #[axum::debug_handler]
 pub async fn online_exchange(
-    State(ctrl): State<crsat::Service>,
+    State(ctrl): State<foreign::crsat::Service>,
     Json(request): Json<web_exchange::OnlineExchangeRequest>,
 ) -> Result<Json<web_exchange::OnlineExchangeResponse>> {
     tracing::debug!("Received request to online exchange");
